new website stuff

+++ data-site/users.content.php
@@ -1,6 +1,10 @@
 <?php
 $db = new Database();
 
+$session = new Session();
+$loggedIn = $session->checkLoginStatus();
+$loggedIn = true;
+
 $sql = 'SELECT `id` FROM `users`;';
 
 $users = $db->object('User');
@@ -10,7 +14,10 @@ $tableStart .= "<th scope=\"col\">ID</th>\n";
 $tableStart .= "<th scope=\"col\">First name</th>\n";
 $tableStart .= "<th scope=\"col\">Last name</th>\n";
 $tableStart .= "<th scope=\"col\">Email</th>\n";
-$tableStart .= "<th scope=\"col\">Actions</th>\n";
+$tableStart .= "<th scope=\"col\">Approved</th>\n";
+if ($loggedIn) { 
+	$tableStart .= "<th scope=\"col\">Actions</th>\n"; 
+} 
 $tableStart .= "</tr>\n";
 
 $tableEnd = "</table>\n";
@@ -23,8 +30,11 @@ foreach ($users as $user) {
 	$tableData .= "<td>{$user->firstname}</td>\n";
 	$tableData .= "<td>{$user->lastname}</td>\n";
 	$tableData .= "<td>{$user->email}</td>\n";
-	$tableData .= "<td><a href=\"user-edit.php?id={$user->id}\" class=\"icon-button\"><i class=\"fas fa-pencil-alt\" role=\"img\" aria-label=\"Edit\"></i></a> ";
-	$tableData .= "<a href=\"user-delete.php?id={$user->id}\" class=\"icon-button\"><i class=\"far fa-trash-alt\" role=\"img\" aria-label=\"Delete\"></i></a></td>\n";
+	$tableData .= "<td>{$user->approved}</td>\n";
+	if ($loggedIn) { 
+		$tableData .= "<td><a href=\"user-edit.php?id={$user->id}\" class=\"icon-button\"><i class=\"fas fa-pencil-alt\" role=\"img\" aria-label=\"Edit\"></i></a> ";
+		$tableData .= "<a href=\"user-delete.php?id={$user->id}\" class=\"icon-button\"><i class=\"far fa-trash-alt\" role=\"img\" aria-label=\"Delete\"></i></a></td>\n";
+	}
 	$tableData .= "</tr>\n";
 }
